Avro

+++ src/Azure.IIoT.OpcUa.Publisher/src/Storage/StateStoreFolderProvider.cs
@@ -10,7 +10,9 @@ namespace Azure.IIoT.OpcUa.Publisher.Storage
     using Microsoft.Extensions.Logging;
     using Microsoft.Extensions.Options;
     using System;
+    using System.Collections.Generic;
     using System.IO;
+    using System.IO.Hashing;
     using System.Linq;
     using System.Threading;
     using System.Threading.Tasks;
@@ -32,24 +34,8 @@ namespace Azure.IIoT.OpcUa.Publisher.Storage
         {
             _serializer = serializer;
             _logger = logger;
+            _hash = new Crc32();
             _stateStoreFolder = options.Value.StateStoreFolderPath;
-        }
-
-        private Stream Open(string key, FileMode mode)
-        {
-            if (_stateStoreFolder == null)
-            {
-                return new MemoryStream();
-            }
-            else
-            {
-                var path = Path.Combine(_stateStoreFolder, typeof(T).Name);
-                if (!Directory.Exists(path))
-                {
-                    Directory.CreateDirectory(path);
-                }
-                return new FileStream(Path.Combine(path, key), mode);
-            }
         }
 
         /// <inheritdoc/>
@@ -60,10 +46,19 @@ namespace Azure.IIoT.OpcUa.Publisher.Storage
                 var fs = Open(key, FileMode.Create);
                 await using (fs.ConfigureAwait(false))
                 {
-                    await _serializer.SerializeAsync(fs, value,
-                        SerializeOption.Indented, ct: ct).ConfigureAwait(false);
+                    var crc = new HashCalculatingStream(fs, _hash);
+                    await using (crc.ConfigureAwait(false))
+                    {
+                        await _serializer.SerializeAsync(crc, value,
+                            SerializeOption.Indented, ct: ct).ConfigureAwait(false);
+                    }
+
+                    _items.TryGetValue(key, out var previous);
+                    var item = new State(crc.HashAndReset, (fs as MemoryStream)?.ToArray());
+
+                    _items.AddOrUpdate(key, item);
+                    return !previous.CheckSum.Span.SequenceEqual(item.CheckSum.Span);
                 }
-                return true;
             }
             catch (Exception ex)
             {
@@ -77,16 +72,23 @@ namespace Azure.IIoT.OpcUa.Publisher.Storage
         {
             try
             {
-                var fs = Open(key, FileMode.OpenOrCreate);
+                var fs = Open(key, FileMode.Open);
                 await using (fs.ConfigureAwait(false))
                 {
-                    return await _serializer.DeserializeAsync<T>(fs,
-                        ct: ct).ConfigureAwait(false);
+                    var crc = new HashCalculatingStream(fs, _hash);
+                    await using (crc.ConfigureAwait(false))
+                    {
+                        var result = await _serializer.DeserializeAsync<T>(crc,
+                            ct: ct).ConfigureAwait(false);
+
+                        _items.AddOrUpdate(key, new State(crc.HashAndReset, default));
+                        return result;
+                    }
                 }
             }
-            catch (Exception ex)
+            catch
             {
-                _logger.LogDebug(ex, "Failed to load state.");
+                _items.AddOrUpdate(key, new State());
                 return default;
             }
         }
@@ -94,29 +96,62 @@ namespace Azure.IIoT.OpcUa.Publisher.Storage
         /// <inheritdoc/>
         public ValueTask RemoveAsync(string key, CancellationToken ct)
         {
-            if (_stateStoreFolder == null)
+            if (_items.Remove(key))
             {
-                return ValueTask.CompletedTask;
-            }
-            try
-            {
-                var path = Path.Combine(_stateStoreFolder, typeof(T).Name);
-
-                File.Delete(Path.Combine(path, key));
-
-                if (!Directory.EnumerateFiles(path).Any())
+                if (_stateStoreFolder == null)
                 {
-                    Directory.Delete(path);
+                    return ValueTask.CompletedTask;
                 }
-            }
-            catch (Exception ex)
-            {
-                _logger.LogDebug(ex, "Failed to remove state.");
+                try
+                {
+                    var path = Path.Combine(_stateStoreFolder, typeof(T).Name);
+
+                    File.Delete(Path.Combine(path, key) + kfileTypeSuffix);
+
+                    if (!Directory.EnumerateFiles(path).Any())
+                    {
+                        Directory.Delete(path);
+                    }
+                }
+                catch (Exception ex)
+                {
+                    _logger.LogDebug(ex, "Failed to remove state.");
+                }
             }
             return ValueTask.CompletedTask;
         }
 
+        /// <summary>
+        /// Open file for key
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="mode"></param>
+        /// <returns></returns>
+        private Stream Open(string key, FileMode mode)
+        {
+            if (_stateStoreFolder == null)
+            {
+                if (mode != FileMode.Create &&
+                    _items.TryGetValue(key, out var previous))
+                {
+                    return new MemoryStream(previous.Buffer.ToArray());
+                }
+                return new MemoryStream();
+            }
+            var path = Path.Combine(_stateStoreFolder, typeof(T).Name);
+            if (!Directory.Exists(path))
+            {
+                Directory.CreateDirectory(path);
+            }
+            return new FileStream(Path.Combine(path, key) + kfileTypeSuffix, mode);
+        }
+
+        private record struct State(ReadOnlyMemory<byte> CheckSum, ReadOnlyMemory<byte> Buffer);
+
+        private const string kfileTypeSuffix = ".json";
+        private readonly Dictionary<string, State> _items = new();
         private readonly ILogger _logger;
+        private readonly Crc32 _hash;
         private readonly IJsonSerializer _serializer;
         private readonly string? _stateStoreFolder;
     }
